fix issues with imports in NI by avoiding nested class
